cleaned up project structure, consolidated apps
