+++ apps/api/views.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import Group
 from django.contrib.auth import get_user_model
 from rest_framework import viewsets
 
-from serializers import UserSerializer, GroupSerializer
+from api.serializers import UserSerializer, GroupSerializer
 
 class UserViewSet(viewsets.ModelViewSet):
     """
